apps/mcp/pyproject.toml
[project]
name = "pantry-pilot-mcp"
version = "0.2.0"
description = "MCP tools for receipt extraction and context retrieval"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "python-dateutil>=2.8.2",
    "minio>=7.2.0",
    "pillow>=10.2.0",
    "httpx>=0.26.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "easyocr>=1.7.0",
]

[tool.uv.sources]
pantry-pilot-shared = { path = "../../packages/shared", editable = true }

[dependency-groups]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*"]